refactor(cli): Add helpers modules for better separation of focus
